==> 1_integrate_data.Rmd <==
---
title: "undergraduate research"
author: "Zhuan_Chen&Boming_Zheng"
date: "2024-12-19"
output: html_document
---

```{r}
library(dplyr)
library(tidyverse)
library(ggplot2)
library(here)
```

# write previous code as a function
```{r}
integrate_ptr <- function (day,Scott = 'Scott_CZA.csv',path = "E:/R_leaning/undergraduate research/PTRDATA") {
    # set the prtdata path
    setwd(path)
    # read ptrnames file
    Scott <- read.csv(Scott)
    chem_names <- read.csv("ptr_chemical_names.csv")
    # got the ptr name by day
    ptrnames <- Scott[str_glue("Day{day}")]
    # attention!!! initial, day is a number, now Day is the integrate file. 
    #the initial data.frame should have the same size of what we bind 'ESC$Range..Raw.Data.'
    Day <- data.frame(m.z = chem_names$m.z)
    # loop for writing column "Range" of ptrdata to integrate file
  for (i in ptrnames[,1]){
    # dynamic parameter transmission for constructing prtdata path
    path_ESC <- str_glue("{i}.tsv")
    # rewrite header to remove 'Day'
    Header_names <- Scott$concentration[ptrnames == i]
    # read ptrdata file
    ESC <- read.table(path_ESC, header = T, sep = '\t')
    Day <- Day %>% bind_cols(!!sym(Header_names) := ESC$Range..Raw.Data.)
  }
    # save integrated file as csv
    res_file <- str_glue("../Day{day}.csv")
    write.csv(Day, res_file, row.names = FALSE)
    return(Day)
}

```



```{r}
day0<-integrate_ptr(0)
day6<-integrate_ptr(6)
day12<-integrate_ptr(12)
```
```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```
