==> 2_Deseq2.Rmd <==
---
title: "2_Deseq2"
author: "Zhuan_Chen&Boming_Zheng"
date: "2024-12-21"
output: html_document
---

```{r}
library(DESeq2)
library(vsn)
library(dplyr)
library(tidyverse)
library(ggplot2)
library(here)
library(purrr)
library(rlang)
library(pheatmap)
```


```{r}
# set work dictionary
setwd("./PTRDATA")
# a file that stores path name of ptrdata
Scott <- read.csv('Scott_CZA.csv')
chem_names <- read.csv("ptr_chemical_names.csv")
```


```{r}
# read integrate file, which from 1_integrate_data
day0 <- read.csv("Day0.csv")
day6 <- read.csv("Day6.csv")
day12 <- read.csv("Day12.csv")
# remove 'AIR' columns
day0 <- day0[, 1:(ncol(day0) - 5)]
day6 <- day6[, 1:(ncol(day6) - 5)]
day12 <- day12[, 1:(ncol(day12) - 5)]
```

# DEseq2
```{r}
# set work dictionary
setwd("./PTRDATA")
# a file that stores path name of ptr_metadata
rawcounts <- day6
metadata <-  read.csv("./ptr_metadata.csv")
```

```{r}
head(rawcounts)
```


```{r}
head(metadata) 
```


```{r}
rawcountsRN <- data.frame(rawcounts, row.names=1) %>% round()
metadataRN <- data.frame(metadata, row.names=1)
```


```{r}
dds <- DESeqDataSetFromMatrix(countData=rawcountsRN, 
                              colData=metadataRN, 
                              # specify the design formula, in this case we are interested in the dex variable in the metadata data frame
                              design=~dex) 
```

# normilization
```{r}
sizeFactors <- estimateSizeFactors(dds) 

# normalized counts based on DESeq2 size factors
normCounts.deseq <- counts(sizeFactors, normalized=TRUE) 

#Let's compare the non to the normalized counts:
head(counts(dds),15)
```

# filter out low value
```{r}
filter <- rowSums(counts(dds) >= 20) >= 3
dds.filtered <- dds[filter, ]
```


```{r}
nrow(counts(dds))-nrow(counts(dds.filtered))
```


```{r}
# run the DESeq pipeline by simply passing the DESeqDataSet object to the DESeq function
dds <- DESeq(dds.filtered) 
```


```{r}
resultsNames(dds)
```


```{r}
# get results for the comparison of treated vs control
res <- results(dds, tidy=TRUE, contrast = c("dex","treated","control")) 

# display the first few rows of the results table
head(res)
```


```{r}
# convert to a tibble
res <- tbl_df(res)
```


```{r}
# add a new column sig to your res data frame, where each entry in sig is TRUE if the corresponding padj value is less than 0.05, and FALSE otherwise.
res <- res %>% mutate(sig=padj<0.05)

# How many of each?
res %>% 
  # grouping the data by the sig column
  group_by(sig) %>% 
  summarize(n=n()) # summarize the data and n=n() function to count the number of rows in each group
```


```{r}
# baseMean as x-axis, log2FoldChange as y-axis, and color by sig
res %>% ggplot(aes(baseMean, log2FoldChange, col=sig)) + 
  # add points
  geom_point() + 
  # log scale the x-axis
  scale_x_log10() + 
  # add a title
  ggtitle("MA plot") 
```


```{r}
# -1*log10() transformation is commonly used to plot p-values
res %>% ggplot(aes(log2FoldChange, -1*log10(pvalue), col=sig)) + 
  geom_point() + 
  ggtitle("Volcano plot")
```


```{r}
rlog.deseq <- rlog(dds.filtered,blind = TRUE)
```

# pca
```{r}
# this time we add the paramter returnData = TRUE, to get the data that was used to create the plot
pcaData <- plotPCA(rlog.deseq, intgroup = c( "dex", "con"), returnData = TRUE)
attr(pcaData, "percentVar")
percentVar <- round(100 * attr(pcaData, "percentVar"))
```
```{r}
ggplot(data = pcaData, mapping = aes(x = PC1, y = PC2, color = con, shape = dex)) +
  geom_point(size = 5) +
  stat_ellipse(aes(group = dex, color = dex),  
  type = "norm",                  
  level = 0.95,                   
  linewidth = 0.5,)+
  xlab(paste0("PC1: ", percentVar[1], "% variance")) +
  ylab(paste0("PC2: ", percentVar[2], "% variance")) 
```
# dist heatmap
```{r}
dist.deseq <- dist(t(assay(rlog.deseq)),"euclidian") 

sampleDistMatrix <- as.matrix(dist.deseq) 

rownames(sampleDistMatrix) <- paste( rlog.deseq$con, rlog.deseq$dex, sep = " _" )

colnames(sampleDistMatrix) <- paste( rlog.deseq$con, rlog.deseq$dex, sep = " _" ) 

pheatmap(sampleDistMatrix,
         clustering_distance_rows = dist.deseq,
         clustering_distance_cols = dist.deseq, 
         clustering_method = "complete",
         angle_col = 45,
         scale = "none") 
```

```{r}
res_highDE <- filter(res,abs(log2FoldChange)>1.5 & padj<0.05)
highestDEname <- res_highDE$row
highestDEexp <- as.data.frame(subset(normCounts.deseq, rownames(normCounts.deseq) %in% highestDEname))
```


```{r}
# Heatmap
df <- as.data.frame(colData(dds)[,c("dex","con")]) 

p <- pheatmap(highestDEexp, 
         scale = "row", 
         clustering_method = "complete", 
         annotation_col=df, 
         show_rownames = T, 
         fontsize_row = 9,
         angle_col = 45,
         cellwidt = 15,
         cellheight = 20,
         color = colorRampPalette(rev(RColorBrewer::brewer.pal(10, "RdBu")))(256),cutree_cols = 3,cutree_rows = 2) 
```







```{r}

```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```
