==> 3_shannon_filter.Rmd <==
---
title: "3_shannon_filter"
author: "Zhuan_Chen&Boming_Zheng"
date: "2025-01-09"
output: html_document
---

```{r}
library(dplyr)
library(rlang)
library(ggplot2)
library(stringr)
library(tidyr)
library(data.table)
```

```{r}
# read integrate file, which from 1_integrate_data
day0 <- read.csv("Day0.csv", row.names = 1)
day6 <- read.csv("Day6.csv", row.names = 1)
day12 <- read.csv("Day12.csv", row.names = 1)
# remove 'AIR' columns
day0 <- day0[, 1:(ncol(day0) - 5)]
day6 <- day6[, 1:(ncol(day6) - 5)]
day12 <- day12[, 1:(ncol(day12) - 5)]
```

```{r}
# # extract data
# chem_con <- day0[1,]
# #chem_name
# chem_name <- row.names(chem_con)
# 
# # p-value
# # replicate of one concentration sample
# len_sample <- ncol(chem_con)/3
# p_value <- c()
# ES_con <-c()
# chem_avg <- c()
# for (i in 1:len_sample) {
#   con_sample <- chem_con[,(3*i-2):(3*i)] %>% as.numeric() %>% t.test()
#   con_p_value <- con_sample$p.value
#   con_avg <- con_sample$estimate
#   con_name <- chem_con %>% colnames()
#   con_name <- strsplit(con_name[3*i],"_")[[1]][1]
#   con <- strsplit(con_name,"ES")[[1]][2]
#   
#   p_value <- c(p_value,con_p_value)
#   ES_con <- c(ES_con,as.numeric(con))
#   chem_avg <- c(chem_avg,con_avg)
# }
# 
# # result
# result <- data.frame(ES_con = ES_con)
# result <- result %>% mutate(p_value,!!sym(chem_name) := chem_avg)
```

```{r}
# for a number, if it is 1-9, convert to 0+number
number_conv <- function(num){
  num <- toString(num)
  if (nchar(num)<2){
    num = str_glue("0{num}")
  }
  return(num)
}
```

```{r}
# function to convert the order of mz to mz
# the column of mz was retrieve from day0 file as default, if the file name is different, please change to the file that have the m/z as rowname.
# !!! mz is character, ID is numeric
ID2mz <- function(chem_ID,day = 0){
  Day <- str_glue("Day{day}.csv")
  Day <- read.csv(Day, row.names = 1)
  Day <- Day[, 1:(ncol(Day) - 5)]
  # extract data
  chem_con <- Day[chem_ID,]
  #mz
  mz <- row.names(chem_con) %>% substr(start = 2,stop = 8)
  return(mz)
}
```

```{r}
# function to convert mz to the order of mz 
# the column of mz was retrieve from day0 file as default, if the file name is different, please change to the file that have the m/z as rowname.
# !!! mz is numeric, ID is numeric
mz2ID <- function(mz,day = 0){
  Day <- str_glue("Day{day}.csv")
  Day <- read.csv(Day, row.names = 1)
  Day <- Day[, 1:(ncol(Day) - 5)]
  ID <- 1
  pattern <- "m.*?\\("
  #re to match mz
  All_mz <- regmatches(row.names(Day), gregexpr(pattern, row.names(Day)))
  for (i in All_mz) {
    len_1 = nchar(i)
    num_i <- substr(i,start = 2, stop = len_1-1) %>% as.numeric()
    if (num_i==mz){
      return(ID)
    }
    ID <- ID+1
  }
}
```

```{r}
# for a chemical, function to get the average of replicates,p-value on the specific day.
day_result <- function(day,chem_ID){
  Day <- str_glue("Day{day}.csv")
  Day <- read.csv(Day, row.names = 1)
  Day <- Day[, 1:(ncol(Day) - 5)]
  # extract data
  chem_con <- Day[chem_ID,]
  #mz
  #mz <- row.names(chem_con) %>% substr(start = 2,stop = 7)
  Day <- number_conv(day)
  chem_name <- str_glue("Day{Day}")  
  
  # p-value
  # replicate of one concentration sample
  len_sample <- ncol(chem_con)/3
  p_value <- c()
  ES_con <-c()
  chem_avg <- c()
  for (i in 1:len_sample) {
    con_sample <- chem_con[,(3*i-2):(3*i)] %>% as.numeric() %>% t.test()
    con_p_value <- con_sample$p.value
    con_avg <- con_sample$estimate
    con_name <- chem_con %>% colnames()
    con_name <- strsplit(con_name[3*i],"_")[[1]][1]
    con <- strsplit(con_name,"ES")[[1]][2]
    
    p_value <- c(p_value,con_p_value)
    ES_con <- c(ES_con,as.numeric(con))
    chem_avg <- c(chem_avg,con_avg)
  }
  
  # result
  result <- data.frame(ES_con = ES_con)
  result <- result %>% mutate(p_value,!!sym(chem_name) := chem_avg)
  # print(chem_name)
  return(result)
}
```

```{r}
# # mutate day by day, remove p-value
# day12_18 <- day_result(0,80) %>% mutate(day_result(6,80)) %>% mutate(day_result(12,80)) %>% select(-p_value)
# 
# # long form convert
# day12_18 <- pivot_longer(day12_18,cols = -ES_con, names_to = "variable", values_to = "value")
```

```{r}
# mutate day by day
con_vs_day <- function(days,chem_ID){
  result <- day_result(days[1],chem_ID)
  days <- days[-1]
  for (i in days) {
  result <- result %>% mutate(day_result(i,chem_ID)) 
  }
  result <- select(result,-p_value)
  return(result)
}
```

```{r}
chem80<-con_vs_day(c(0,6,12),80)
```

```{r}
# long form convert
day_80 <- pivot_longer(con_vs_day(c(0,6,12),80),cols = -ES_con, names_to = "variable", values_to = "value")
```

```{r}
ggplot(day_80,aes(x = ES_con,y = value,color = variable))+
  geom_point()+
  geom_line()
```

```{r}
# change to order by columns
order_of_con_vs_day_c <- function(con_vs_day) {
  result <- data.frame(ES_con = con_vs_day[,1])
  data <- con_vs_day[,-1]
  for (i in 1:ncol(data)) {
    result <- result %>% mutate(!!sym(colnames(data)[i]) := order(data[,i]))
  }
  rownames(result) <- con_vs_day[,1]
  return(result)
}

```

```{r}
order_of_con_vs_day_c(chem80)[-1]
```

```{r}
# change to order by rows
order_of_con_vs_day_r <- function(con_vs_day) {
  result <- transpose(con_vs_day[-1])
  data <- result
  for (i in 1:ncol(data)) {
    result <- result %>% mutate(!!sym(colnames(data)[i]) := order(data[,i]))
  }
  result <- transpose(result)
  result <- data.frame(con_vs_day[,1],result)
  result <- transpose(result)
  rownames(result) <- colnames(con_vs_day)
  colnames(result) <- result[1,]
  return(result)
}

```

```{r}
order_of_con_vs_day_r(chem80)[-1,]
```

```{r}
# eq_col: the index of columns with the same order of first column;
#         if count = True , return the count of columns that with the same order of first column.
eq_cols <- function(order_of_con_vs_day_r,count = FALSE){
  # if ncol(oc)==1, oc[-1,] would change from 'data.frame' to 'numeric'. ncol() do not known 'numeric', (return 0,even 'numeric' has elements)
  oc <- order_of_con_vs_day_r
  
  if (ncol(oc)==1){
      if (count) {
      a = 0
      for (i in 1:ncol(oc)){
        if (all(oc[1]==oc[i])){
          a <- a +1
        }
      }
      return(a)
    }
    oc_eq <- c(1)
    return(oc_eq)
  
  }
  
  oc <- oc[-1,]
  
  if (count) {
    a = 0
    for (i in 1:ncol(oc)){
      if (all(oc[1]==oc[i])){
        a <- a +1
      }
    }
    return(a)
  }
  oc_eq <- c()
  for (i in 1:ncol(oc)){
    if (all(oc[1]==oc[i])){
      oc_eq <- c(oc_eq,i)
    }
  }
  return(oc_eq)
}
```

```{r}
# frequency: the percentage of all type order.
frequence <- function(order_of_con_vs_day_r){
  oc <- order_of_con_vs_day_r
  t_count <- ncol(oc)
  result <- c()
  
  eq <- eq_cols(oc)
  c_eq <- eq_cols(oc,count = T)
  result <- c(result,c_eq/t_count)
  
  while(ncol(oc[-eq])>0){
    oc <- oc[-eq]
    eq <- eq_cols(oc)
    c_eq <- eq_cols(oc,count = T)
    result <- c(result,c_eq/t_count)
  }
  return(result)
}
```


```{r}
freq <- frequence(order_of_con_vs_day_r(con_vs_day(c(0,6,12),60)))
freq
```

```{r}
shannon_index <- function(frequence){
  k <- 0
  for (i in frequence){
    k <- k-i*log2(i)
  }
  k
}
```

```{r}
shannon_index(freq)
```
```{r}
chaos <- function(day,index){
    chem <- con_vs_day(day,index)
    order_r <- order_of_con_vs_day_r(chem)
    freq <- frequence(order_r)
    sh_index <- shannon_index(freq)
    return(sh_index)
}
```

```{r}
chaos_list <- function(day = c(0,6,12),start = 1,end = 318){
  result <- data.frame(index = start:end)
  chaos_res <- c()
  for (i in start:end){
    # error at t.test, if the data are constant, t.test,return error
    sh_index <- tryCatch(expr = chaos(day,i),error = function(e){return(NaN)})
    chaos_res <- c(chaos_res,sh_index)
  }
  result <- result %>% mutate(shannon = chaos_res)
  return(result)
}
```


```{r}
shannon_result <- chaos_list()
shannon_result
```


```{r}
rownames(shannon_result) <- rownames(day0)
shannon_result
```


```{r}
barplot(shannon_result$shannon,
        main = "shannon_index",
        xlab = "m/z",
        ylab = "shannon"
        )
```


```{r}
ggplot(shannon_result, aes(x = shannon)) +
  geom_bar(fill = "skyblue") +
  geom_text(stat = "count", aes(label = after_stat(count)), vjust = 1, size = 3.5) +
  labs(title = "Frequencies For shannon", x = "Shannon", y = "Frequence") +
  theme_minimal()
```


```{r}
cza_plot <- function(index,day = c(0,6,12),save_fig = FALSE,by_mz = F){
  # convert mz to the order of mz (index)
  if (by_mz){
    index = mz2ID(index)
  }
  # get shannon index
  shannon = round(shannon_result$shannon[index],digits = 2)
  # long form convert
  long_cov <- pivot_longer(con_vs_day(day,index),cols = -ES_con, names_to = "variable", values_to = "intensity")
  
  fig <- ggplot(long_cov,
                aes(x = ES_con,y = intensity,
                    color = variable))+
    geom_point(size = 3)+
    geom_line()+
    labs(
      title = str_glue("m/z=",ID2mz(index)),
      subtitle = str_glue("Shannon=",shannon)
    )+
    theme(title = element_text(size =20),
          axis.title = element_text(size = 20),
          axis.text = element_text(size = 15),
          legend.text = element_text(size = 20),
          )
  if (save_fig){
    ggsave(str_glue("./fig/",toString(ID2mz(index)),'.png'),fig,limitsize = F,width = 4,height = 3)
  }
  plot(fig)
}
```


```{r}
cza_plot(149.096,save_fig = T,by_mz = T)
```


```{r}
sessionInfo()
```











```{r}

```






```{r}
```


```{r}
```


```{r}
```


```{r}
```
