==> 4_spyder_for_mVOC.Rmd <==
---
title: "4_spyder_for_mVOC"
author: "Zhuan_Chen&Boming_Zheng"
date: "2024-12-21"
output: html_document
---
```{r}
# 处理html
library(httr)
# 发送亲求
library(rvest)
# taxname to taxid or search 'class,phylum...' by taxname
library(taxize)
```


```{r}
# 构造POST请求
url <- "https://bioinformatics.charite.de/mvoc/index.php?site=ergebnis"
form_data <- list(
  compound_id = 1068,          # 化合物ID
  name = "",                 # 化合物名称
  formula = "",              # 化学式
  weightvon = '',            # 分子量范围（起始）
  weightbis = '',            # 分子量范围（结束）
  logPvon = "",              # logP范围（起始）
  logPbis = "",              # logP范围（结束）
  classification = "all",    # 化合物分类
  species_bacteria = "all",  # 细菌物种
  species_fungi = "all",     # 真菌物种
  organism = "Please choose the species first",
  B3 = "Search"              # 搜索按钮标识
)

# 设置请求头
headers <- c(
  "User-Agent" = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/132.0.0.0 Safari/537.36 Edg/132.0.0.0",
  "Referer" = "https://bioinformatics.charite.de/mvoc/index.php?site=mvocsearch",
  "Accept" = "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7"
)
```


```{r}
# 发送POST请求
response <- POST(url, body = form_data, add_headers(.headers = headers))

# 检查响应状态码
stop_for_status(response)

# 解析返回的HTML内容
html <- content(response, "text") %>%
  read_html()
```


```{r}
# 提取物种信息（假设物种信息出现在特定的HTML节点中）
# 请根据实际返回的HTML结构调整选择器
species <- html %>%
  html_nodes("body div.shadow_home table.ref")%>%.[1]%>% html_nodes("tr")%>% .[-1] %>% html_nodes('td a.table_left') %>% 
# 选择器,根据html对象的类型和属性定位数据
# unique(),降重
 html_text() %>% unique()

# 打印结果
print(species)
```


```{r}
tax_ids <- get_ids(species,db = 'ncbi')
tax_ids
```


```{r}
tax_ids <- tax_ids$ncbi
tax_ids
```


```{r}
ncbi_id <- c()
for (i in tax_ids) {
  ncbi_id <- c(ncbi_id,as.numeric(i))
}
ncbi_id
```


```{r}
phylum_name <- tax_name(species,get = "phylum",db = 'ncbi')
phylum_name
```

```{r}
result <- data.frame(id = ncbi_id) %>% mutate(phylum_name)
```


```{r}
write.table(result, file = "result.csv", sep = ",", row.names = FALSE, col.names = TRUE) 

write.table(ncbi_id, file = "ncbi_id.txt", sep = ",", row.names = FALSE, col.names = TRUE) 
```


```{r}

```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```


```{r}
```
